Separate actual source code from metadata information
This way you can install only the actual source code into Blender by copy-pasting that folder.
If there are a few README files in the add-ons folder of Blender that is not so bad, but it'll get progressively worse if I start adding some images for the readme as well as automated tests and a CMake installation file.