Use unittest's own assertion function rather than assert keyword
This way, the error messages you see when the test fails are much more clear. It will show the evaluated values that it's comparing rather than just an AssertionError with a message and a traceback to the line of code. Assertion methods with list comparison can even show which elements of the list are not equal.
I've also transformed all of the len(x) == 0 checks to check if lists are equal to []. This way, the list equal assertion will show what the content is of the list if it's not empty.