Move use_mesh_modifiers property to the fixture
We're not ever going to test this. Might as well define it here.
We can't really test this because it involves only integration with Blender, which needs to be mocked out. We could test that the correct functions are being called, but all that really does is duplicate the code.