Correct order of applying global transformation
Yeah, the order was wrong. This reveals a flaw in the way I wrote that test. I didn't actually know what the correct order would be. I just tested that manually and determined the correct order that way. Now it turns out that this manual test wasn't conclusive. These needed to be swapped.