Separate parent transformation from child when saving
In 3MF, transformations from components are multiplied with the transformations of their parent object resources. Not so in Blender. In Blender, each object has their own transformation which is completely separate of the transformation of any anchestral objects. When importing we were doing this properly: We were multiplying the transformations with the parent transformations and applying the multiplied transformation to each object. So when exporting we'd like to get the original transformation back. To do that, we multiply with the inverse of the parent's transformation, sort of like a division, although with matrix multiplication not being commutative the order is very important here so it's not quite as simple.
Matrix inverse is not always possible if one of the axes is scaled down to 0. We don't know what the original scale was then; there is just not enough information any more because anything times 0 is 0. Instead we'll use a pseudoinverse which doesn't fail in those cases by assuming a very small epsilon. This epsilon doesn't matter in the final result though, because due to the parent's transformation it'll get multiplied to 0 in the final coordinates. However that does mean that the transformations are slightly different for the child objects (components) in a read-write loop, and if we make integration tests for that we'll need to keep that in mind.