Implement reading metadata tags from the archive
This is metadata in the scene. Blender has no real structure to save this. I suppose we'll just have to add a text file to store the stuff, preserving it for when we write again. We'll store what we can on the objects themselves for Blender to display. But maybe that will only be the title or something.