Use new metadata class to store metadata
This makes a lot of things much simpler. The whole resolution can be simplified since it doesn't need to store multiple entries for each metadata key and resolve them on the spot or anything. And instead of muddying up the main import class we're doing that separately in a class that can get its own set of tests, making those more overseeable as well.