Write a text file to the scene containing metadata
The metadata is serialised as JSON and stored in the Blender context. It'll be hidden, so only the Blender API can access it. I'm prefacing the ID of this file with '3mf' so I'm pretty sure it won't cause accidental collisions.
This way the exporter can read the metadata that was stored previously and export it again.