Reload previously-loaded metadata from other 3MF files
If you're loading multiple 3MF files in one scene, there can be metadata which is the same and some metadata which is different. This routine will cause the 3MF metadata to be combined and re-saved. If there are conflicts of metadata, those will be erased. If there is metadata which is the same, those are kept.