Allow metadata reading function to be re-used for metadata groups
This means that it must accept any XML node (not just the root) and that it must return a new Metadata entry for those.
In case of the metadata in the root, we do want to end up with just 1 metadata group, not 1 for every document. So those must be combined. To that end we must also be able to provide our own metadata group to combine with to the function.