Move build_metadata into Metadata class
We're going to need to retrieve the metadata from Blender objects in the importer as well. The importer needs to do both, in order to combine metadata from multiple files. So to facilitate the exporter also using the retrieval function, we're moving these to a central location: in the Metadata class itself.