Retrieve pre-existing metadata and combine with imported
If you've loaded a different 3MF file previously, the metadata needs to be combined.
The title is now never applied though, since Blender always gives a scene a title so that will always be there and (almost) always be different than the title in the 3MF document. Need to do something about that.