Don't restore title from the scene
If we do, we'll always end up with the scene title 'Scene', because that's how the scene originally gets created by Blender. If the 3MF document contains any other title, that will be in conflict with the original metadata which contains the title 'Scene', and the Metadata object won't contain any entry for the title. Then we don't write the new title to the scene title. If the 3MF document contains the same title, the metadata won't be in conflict but since the title is 'Scene' we'd write that title which doesn't change the scene's title then.
Now we'll delete the original title from the metadata. If the 3MF document contains a title, that one will be the new title for the scene. If it doesn't, we'll keep the original title because the Metadata instance won't override the title of the scene then.