Store part numbers as metadata entries as well
The spec says that this metadata is special because it should be retained regardless of which operations are done on the data, as long as the data is based on an original mesh or item with a part number, the transformation should have that same part number. Blender can't guarantee that. It'll maintain the properties for most things, but as soon as it creates new meshes (e.g. by splitting up meshes) or deletes meshes, this metadata is no longer retained.
But for the most part it's fine. I'm just handling it as any other metadata which also gets retained through these transformations.