Write object resource metadata, not item metadata
The object resource metadata is saved to the mesh, not to the object. So retrieve the metadata from the mesh. Also, only write it if there is a mesh.
We need to consider using an empty mesh just to store metadata, perhaps, in the import stage. I'll write that down.