Move parsing of rels file to separate annotations class
Let's move the parsing, serialising, deserialising and writing of rels and content types to a separate class.
We can then test this class separately, resulting in a bit easier to understand tests. And the code itself should be easier to find.
But most importantly, we can then re-use the same serialisation code, and the code for serialisation/deserialisation will be very close together.