Move mocking of bpy package to __init__
For two good reasons. One: It's annoying to repeat all that magic code for every test case we add. Two: The actual mock used in the code under test may be different from the mock used in the test itself. This is because the code under test gets its module imported when the code under test is first imported, which is in my case for the Import tests. But the test itself may get its import done when sys.modules contains a wholly new mock created in the module that the test is contained in. So we can't compare data from the mock in the code under test because that mock is not in sys.modules any more.