Simplify serialisation of annotations
We were storing a whole dictionary as the single element in the list, just to get an extra key in that would indicate the annotation type. Instead, we'll now smash that dictionary flat and just put the type as an extra field in the dictionary holding the fields of the annotation.