Handle conflicting MustPreserve files
These are files which must be preserved according to multiple archives, but the contents of the files are different while the file name is the same. In the resulting export, there can only be one output file with that file name, so it's ambiguous which archive's content would be preserved. In that case, the intent is to preserve neither. For now we'll just store a marker file in the Blender scene so that the importer can see that there are files in conflict here. The exporter must ignore those files.