Add function to parse material resources from 3MF documents
This does just the parsing. It doesn't create actual materials yet in Blender. It'll only create the materials that are actually used in the scene, and we don't know yet at this stage which materials will be used, so we can't actually add the materials to Blender yet. That needs to happen when the items are built.