Only store one material per triangle
This is due to a limitation of Blender. It turns out that Blender can only assign materials per face. It can't assign a different material for each vertex and blend between them.
The 3MF specification says that if the consumer doesn't support blending between materials, it should only use the one from the first vertex. So there you go.
Weird that Blender doesn't support blending, considering the name ;)