Correct colour space for imported material diffuse colours
Blender works with linear radiant colours: If a surface emits twice as many photons, the colour value will be twice as high. However 3MF works with sRGB colours. These colours are optimised for a display and have a limited range. To use the sRGB colour we can create a render node for the material that uses a PrincipledBSDF material wrapper to simulate a material with the same reflectance as the sRGB display.
I took this from the example of the OBJ io add-on of Blender and the example provided in Blender's 2.80 release notes (https://wiki.blender.org/wiki/Reference/Release_Notes/2.80/Python_API/Modules).