Mock the PrincipledBSDFWrapper with a custom mock
This mock transparently wraps the material that the BSDFWrapper would wrap. The base colour is the diffuse colour of the material, and the alpha is the alpha channel of the diffuse colour. Nothing gets transformed, so it's possible that the colours go outside of the colour gamut of sRGB in this case (but then again, maybe the consumer needs to handle that case well too).
This is better than using a MagicMock because the MagicMock would return MagicMocks for all of the colour values, which our code doesn't deal with well. It's easier to mock the entire class than to mock the results from the MagicMock.