Zoom camera to imported objects automatically
This is necessary because 3MF meshes are often meant for 3D printed models, and have correct units for those models. Those models are usually in millimetres, so they will appear very small in Blender's default scene set-up which uses metres as grid and oversees a scene of several dozen metres.

Fixes #28.