Limit line length to 120 characters
This is the limit imposed by Blender's code style.
I deeply believe that arbitrary limits like this only serve to reduce readability. It reduces the flexibility we're allowed to line up code together to improve readability with semi-repeating code (e.g. two calls with the same parameters). Especially when combined with PEP-8, which doesn't allow aligning comments on separate lines.