Fix running tests on Windows
In Python the resources are destroyed when the variable goes out of scope, but the variable doesn't go out of scope until the end of the function. So when the file is being deleted it is still open. Unix-like systems don't care. They delete the file but keep the file node until the process releases it. But on Windows this causes a PermissionError that the file is still in use.