Use an ID name that is more friendly to scripting
This is how one would be able to access this operator from a script. However in the Python language, names starting with a number are not valid, so calling bpy.ops.import_mesh.3mf() would result in a syntax error. You can now call bpy.ops.import_mesh.threemf(filepath='...') to import a mesh from a script, or similarly when exporting.

Fixes #34.